refactor: move node builder fns into nodesBuilder file
